refactor(mobile): extract activity participation hook
