+++ apps/mobile/src/lib/use-activity-detail-screen.ts
@@ -131,6 +131,7 @@ export function useActivityDetailScreen({ activityId, onActivityLoaded }: Props)
 
   return {
     activity,
+    setActivity,
     viewerUserId,
     loading,
     isEditing,
